+++ bench_state.go
@@ -79,13 +79,14 @@ func (s *benchmarkState) recordLatency(d time.Duration) {
 	s.statter.Timing("latency", d)
 }
 
-func (s *benchmarkState) printLatencies(out output) {
-	// TODO JSON output?
+// Returns a mapping of quantiles to latency values
+func (s *benchmarkState) getLatencies() map[float64]time.Duration {
 	sort.Sort(byDuration(s.latencies))
-	out.Printf("Latencies:\n")
-	for _, quantile := range []float64{0.5, 0.9, 0.95, 0.99, 0.999, 0.9995, 1.0} {
-		out.Printf("  %.4f: %v\n", quantile, s.getQuantile(quantile))
+	latencyValues := make(map[float64]time.Duration, len(_quantiles))
+	for _, quantile := range _quantiles {
+		latencyValues[quantile] = s.getQuantile(quantile)
 	}
+	return latencyValues
 }
 
 func (s *benchmarkState) printErrors(out output) {