+++ options.go
@@ -109,6 +109,7 @@ type BenchmarkOptions struct {
 	// Benchmark metrics can optionally be reported via statsd.
 	StatsdHostPort string `long:"statsd" description:"Optional host:port of a StatsD server to report metrics"`
 	PerPeerStats   bool   `long:"per-peer-stats" description:"Whether to emit stats by peer rather than aggregated"`
+	Format         string `long:"format" description:"Prints benchmark output in either text or JSON format. Default is text."`
 }
 
 func newOptions() *Options {
@@ -172,10 +173,6 @@ func (s *stringAlias) UnmarshalFlag(value string) error {
 	}
 	*s.dest = value
 	return nil
-}
-
-func (s *stringAlias) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	return unmarshal(s.dest)
 }
 
 func setEncodingOptions(opts *Options) {